refactor(api): move storage module under util and simplify enqueuer wiring
- Relocate storage.rs to util/storage.rs to group it with distance/error utilities.
- Update all import sites in GraphQL mutations and objects.
- Extract TrackPointEnqueuer construction into build_track_point_enqueuer().
- Drop redundant Arc<TrackPointEnqueuer> wrapper; rely on async-graphql data storage.
