+++ apps/api/src/graphql/mutation/user.rs
@@ -4,7 +4,7 @@ use sea_orm::ActiveModelTrait;
 use sea_orm::ActiveValue::Set;
 
 use crate::graphql::object::user::User;
-use crate::storage::upload_avatar;
+use crate::util::storage::upload_avatar;
 use crate::{entity::user, graphql::guard::AuthGuard};
 
 #[derive(Default, Debug)]
