+++ apps/api/src/lib.rs
@@ -2,5 +2,4 @@ pub mod auth;
 pub mod entity;
 pub mod graphql;
 pub mod queue;
-pub mod storage;
 pub mod util;
