+++ apps/api/src/util/mod.rs
@@ -1,2 +1,3 @@
 pub mod distance;
 pub mod error;
+pub mod storage;
